+++ src/main/kotlin/me/gurkz/superdupermod/SuperDuperMod.kt
@@ -8,6 +8,7 @@
 
 package me.gurkz.superdupermod
 
+import me.gurkz.superdupermod.command.ServerSayCommand
 import me.gurkz.superdupermod.command.SmiteCommand
 import me.gurkz.superdupermod.item.ModItems
 import me.gurkz.superdupermod.network.RespawnPlayer
@@ -33,6 +34,7 @@ object SuperDuperMod : ModInitializer {
         LOGGER.info("hi from super duper mod version $VERSION")
         registerSuperDuperModCommand()
         SmiteCommand.register()
+        ServerSayCommand.register()
 
         ModItems.initialise()