+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - respawn player command ``/superdupermod respawn <player>``
+- server say command ``/superdupermod server-say <message>``
 
 ### Changed